Добавить изменения checkWeightPokemon() тест закончен и работает
