+++ src/test/java/co/pokeapi/config/EndPoints.java
@@ -1,4 +1,4 @@
-package co.pokeapi.dataTests;
+package co.pokeapi.config;
 
 public class EndPoints {
     public static final String POKEMON = "pokemon/";
