+++ src/test/java/co/pokeapi/dataTests/Root.java
@@ -1,0 +1,19 @@
+package co.pokeapi.dataTests;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Root {
+    private Integer weight;
+
+    public Root(){
+    }
+
+    public Root(Integer weight) {
+        this.weight = weight;
+    }
+
+    public Integer getWeight() {
+        return weight;
+    }
+}
