+++ src/test/java/co/pokeapi/steps/PokemonSteps.java
@@ -1,36 +1,31 @@
 package co.pokeapi.steps;
 
 import co.pokeapi.dataTests.Ability;
-import co.pokeapi.dataTests.Pokemon;
+import co.pokeapi.dataTests.Root;
 import io.qameta.allure.Step;
-import io.restassured.response.Response;
-import io.restassured.response.ValidatableResponse;
 
 import java.util.List;
 
-import static co.pokeapi.dataTests.EndPoints.POKEMON;
+import static co.pokeapi.config.EndPoints.POKEMON;
 import static io.restassured.RestAssured.given;
 
 public class PokemonSteps extends RestClient {
 
-//    @Step("Получаем  параметр Вес")
-//    public List<Pokemon> getWeight(String namePokemon) {
-//        List<Pokemon> pokemon = given()
-//                .when()
-//                .spec(getDefaultRequestSpec())
-//                .get(POKEMON + "namePokemon")
-//                .then()
-//                .extract()
-//                .body()
-//                .jsonPath()
-//                .getList("abilities.ability", Pokemon.class);
-//
-//        return pokemon;
-//    }
+    @Step("Получаем  параметр Вес покемона")
+    public int getWeightPokemon(String namePokemon) {
+        return given()
+                .when()
+                .spec(getDefaultRequestSpec())
+                .get(POKEMON+namePokemon)
+                .then()
+                .extract()
+                .body()
+                .as(Root.class).getWeight();
+    }
 
     @Step("Получаем список способностей покемона")
     public List<Ability> getAbility(String namePokemon) {
-        List<Ability> ability = given()
+        return given()
                 .when()
                 .spec(getDefaultRequestSpec())
                 .get(POKEMON+"rattata")
@@ -39,7 +34,5 @@ public class PokemonSteps extends RestClient {
                 .body()
                 .jsonPath()
                 .getList("abilities.ability", Ability.class);
-
-        return ability;
     }
 }
